requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==25.1.0
bcrypt==4.2.1
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-users==14.0.1
Flask==3.1.0
flask-cors==5.0.1
Flask-DotEnv==0.1.2
Flask-MySQLdb==2.0.0
h11==0.14.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.1
makefun==1.15.6
MarkupSafe==3.0.2
motor==3.7.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.3
outcome==1.3.0.post0
packaging==24.2
passlib==1.7.4
pwdlib==0.2.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.1
PyMySQL==1.1.1
PySocks==1.7.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.45.3
trio==0.29.0
trio-websocket==0.12.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
uvicorn==0.34.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0
Werkzeug==3.1.3
wsproto==1.2.0
